pyproject.toml
[tool.poetry]
name = "ml4paleo"
version = "0.1.0"
description = "Deep learning computer vision pipelines for paleontological 3D reconstruction"
authors = ["Jordan Matelsky <j6k4m8@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.1"
Pillow = "^9.4.0"
psutil = "^5.9.4"
torch = "^1.13.1"
torchvision = "^0.14.1"
ipykernel = "^6.21.1"
matplotlib = "^3.6.3"
zarr = "^2.13.6"
tqdm = "^4.64.1"
Flask = "^2.2.2"
scikit-learn = "^1.2.1"
scikit-image = "^0.19.3"
marshmallow = "^3.19.0"
jque = "^0.1.3"
intern = "^1.4.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
jupyter = "^1.0.0"
ipython = "^8.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
